vfs: move the secure path resolver into vfs/secure_open.c
Relocate the race-safe resolver and its policy gates out of syscall.c
into vfs/secure_open.c, and give them the vfs_* public names that the
mainline code will use going forward:

  secure_relative_open    -> vfs_resolve_open
  secure_relative_open_at -> vfs_resolve_open_at
  secure_relpath_active   -> vfs_relpath_active
  symlink_optout_allowed  -> vfs_symlink_optout_allowed

secure_walk_at stays file-local.  The function bodies are unchanged; the
call sites across receiver/sender/generator/flist/util1/clientserver/
main/options and the test harnesses are updated to the new names, and
the four entry points are declared in vfs/vfs.h.

The resolver's only consumer of am_chrooted in syscall.c left with it,
so the now-dead `am_chrooted` is dropped from the do_*_at wrappers'
local externs (and the file-scope extern), which is otherwise unused.
No behavior change.
