+++ util1.c
@@ -393,38 +393,14 @@ int copy_file(const char *source, const char *dest, int tmpfilefd, mode_t mode)
 	OFF_T prealloc_len = 0, offset = 0;
 
 	/* For any hardened (non-chrooted) receiver, route the source open through
-	 * secure_relative_open so a parent-symlink on the source path (e.g.
+	 * vfs_resolve_open so a parent-symlink on the source path (e.g.
 	 * --copy-dest=cd where cd is a symlink to an outside directory) cannot
 	 * redirect the read to a file the attacker should not see.  Plain
 	 * do_open_nofollow only refuses a final-component symlink; parents are
-	 * still followed.  An ABSOLUTE source is an operator basis (e.g. an absolute
-	 * --copy-dest): confine its parents via the ownership walk -- a foreign-owned
-	 * parent symlink is refused, the operator's own dirs/uid0/euid symlinks
-	 * followed -- so a flipped parent can't redirect the basis read out of tree.
-	 * operator_path_resolve is set only across the walk (so module-exclude is
-	 * enforced) and restored, leaving the caller's value for the dest side -- this
-	 * is why confining the source here does not re-open the copy_xattrs dest
-	 * race the way wrapping the whole copy_altdest_file would. */
-	if (secure_relpath_active() && source && *source && source[0] != '/')
-		ifd = secure_relative_open(NULL, source, O_RDONLY | O_NOFOLLOW, 0);
-#if defined AT_FDCWD && defined O_NOFOLLOW && defined O_DIRECTORY
-	else if (secure_relpath_active() && source && source[0] == '/'
-	      && !symlink_optout_allowed()) {
-		int save = operator_path_resolve, dfd, e;
-		const char *leaf;
-		operator_path_resolve = 1;
-		dfd = owner_walk_parent(source, &leaf);
-		operator_path_resolve = save;
-		if (dfd < 0)
-			ifd = -1;
-		else {
-			ifd = openat(dfd, leaf, O_RDONLY | O_NOFOLLOW);
-			e = errno;
-			close(dfd);
-			errno = e;
-		}
-	}
-#endif
+	 * still followed.  (An absolute source is operator-trusted -- e.g. an
+	 * absolutized basis dir -- and uses do_open_nofollow.) */
+	if (vfs_relpath_active() && source && *source && source[0] != '/')
+		ifd = vfs_resolve_open(NULL, source, O_RDONLY | O_NOFOLLOW, 0);
 	else
 		ifd = do_open_nofollow(source, O_RDONLY);
 	if (ifd < 0) {
@@ -1317,7 +1293,7 @@ int change_dir(const char *dir, int set_path_only)
 			 * target -- otherwise CWD escapes the module and
 			 * every subsequent path-relative syscall (open,
 			 * chmod, lchown, ...) inherits the escape, which
-			 * defeats secure_relative_open's RESOLVE_BENEATH
+			 * defeats vfs_resolve_open's RESOLVE_BENEATH
 			 * anchor and re-opens the CVE-2026-29518 class of
 			 * symlink TOCTOU attacks. Use the secure resolver
 			 * to get a confined dirfd, then fchdir() to it.
@@ -1344,7 +1320,7 @@ int change_dir(const char *dir, int set_path_only)
 					prefix[save_dir_len] = '\0';
 					basedir = prefix;
 				}
-				dfd = secure_relative_open(basedir, dir,
+				dfd = vfs_resolve_open(basedir, dir,
 					O_RDONLY | O_DIRECTORY, 0);
 				if (dfd < 0) {
 					chdir_failed = 1;
